requirements.txt
# Core LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph-checkpoint-sqlite>=1.0.0

# LLM Provider - OpenAI
openai>=1.0.0
langchain-openai>=0.0.2

# Telegram Bot
python-telegram-bot>=21.0

# API Server
fastapi>=0.104.0
uvicorn>=0.24.0

# Configuration
python-dotenv>=1.0.0

# Storage & Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# RAG & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# HTTP & Web Scraping
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Google Calendar (for later phases)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0

# iCloud Calendar (for later phases)
pyicloud>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

pydantic>=2.7.0

aiosqlite==0.21.0

qdrant-client>=1.3.0